refactor(basic): 계산 관련 로직 분리
App.tsx에 혼재되어 있던 계산 관련 함수들을 utils/calculators.ts 파일로 분리

- getMaxApplicableDiscount
- calculateItemTotal
- calculateCartTotal
